Add files via upload
the file consists of Compiled Python code, which can import as pairedttest, the package has a method as tscore() which will accepts three arguments as dataset, the metaname in the csv 'X' sample, and 'Y' sample, the data passed to the method, it will removed the rows if any row of columns is null, empty. the Data for both column should be numerical. Once all the data is uploaded, the method calculate the Paired T Test score and return as float64 value or int based on the output
import pandas as pd
import pairedttest as ptt
ds = pd.read_csv('XYSamples.csv')
t_score=ptt.tscore(ds,'X','Y')
print(t_score)